pyproject.toml
[project]
name = "rag-crawler"
version = "0.1.0"
description = "RAG Data Ingestion Crawler using Crawl4AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.4.0",
    "psutil>=5.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "docling>=2.55.0",
    "asyncpg>=0.30.0",
    "transformers>=4.40.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    # LangGraph Agentic RAG
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    # Retrieval & Reranking
    "flashrank>=0.2.0",
    "sentence-transformers>=3.0.0",
    # UI
    "streamlit>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
rag-crawler = "rag_crawler.main:main"
rag-ingest = "rag_crawler.ingestion.ingest:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_crawler"]

[tool.hatch.build.targets.sdist]
include = ["src/rag_crawler"]

[tool.hatch.build]
sources = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["rag_crawler"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["src/rag_crawler/tests"]
asyncio_mode = "auto"
